fix(android): add up saturated edge insets without wrapping
`EdgePadding` carries arbitrary JavaScript numbers, and a finite value past the
`Int` range lands on `Int.MAX_VALUE` in `roundToInt`. Two such insets on one
axis wrapped their sum to -2, so the viewport looked *larger* than it is, the
guard for insets that fill it never fired, and the fit ran on bounds clamped to
the whole Mercator range instead of being skipped.

The aggregates are `Long` now, which is enough: an absurd inset makes the
content box negative and falls into the guard that was meant to catch it.

Caught by CodeRabbit on #165.

+++ package/android/src/main/java/com/margelo/nitro/nitromaps/EdgePaddingPixels.kt
@@ -13,7 +13,9 @@ internal data class EdgePaddingPixels(
   val bottom: Int,
   val left: Int,
 ) {
-  val horizontal: Int get() = left + right
-  val vertical: Int get() = top + bottom
+  // Long, because two insets that saturated at `Int.MAX_VALUE` on the way in would wrap
+  // their sum negative and read back as a viewport with room to spare.
+  val horizontal: Long get() = left.toLong() + right.toLong()
+  val vertical: Long get() = top.toLong() + bottom.toLong()
   val isEmpty: Boolean get() = top == 0 && right == 0 && bottom == 0 && left == 0
 }
